+++ src/main/java/dev/willyelton/crystal_tools/client/gui/CrystalBackpackScreen.java
@@ -63,7 +63,6 @@ public class CrystalBackpackScreen extends ScrollableContainerScreen<CrystalBack
 
     @Override
     public void render(@NotNull GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick) {
-        this.renderBackground(guiGraphics, mouseX, mouseY, partialTick);
         super.render(guiGraphics, mouseX, mouseY, partialTick);
         this.renderTooltip(guiGraphics, mouseX, mouseY);
     }