+++ src/main/java/dev/willyelton/crystal_tools/common/levelable/tool/CrystalApple.java
@@ -1,11 +1,17 @@
 package dev.willyelton.crystal_tools.common.levelable.tool;
 
 import dev.willyelton.crystal_tools.common.components.DataComponents;
+import dev.willyelton.crystal_tools.common.components.EffectData;
 import dev.willyelton.crystal_tools.common.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.utils.ToolUtils;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Holder;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
@@ -13,12 +19,14 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.food.FoodProperties;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.UseAnim;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
-import org.jetbrains.annotations.Nullable;
 
-// TODO: Food things seem to be data components now, might need / want to do that?
-// Might be fine, ItemStack#getFoodProperties only calls my method
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+
 public class CrystalApple extends LevelableTool {
     private static final int BASE_NUTRITION = 2;
     private static final float BASE_SATURATION = 0.4F;
@@ -29,7 +37,7 @@ public class CrystalApple extends LevelableTool {
     }
 
     @Override
-    public FoodProperties getFoodProperties(ItemStack stack, @Nullable LivingEntity entity) {
+    public FoodProperties getFoodProperties(ItemStack stack, LivingEntity entity) {
         return getFoodPropertiesFromNBT(stack);
     }
 
@@ -45,21 +53,18 @@ public class CrystalApple extends LevelableTool {
         if (alwaysEat) builder.alwaysEdible();
 
         // TODO: Effect refactor
-//        if (stack.getTag() != null) {
-//            for (int i = 1; i < 34; i++) {
-//                if (stack.getTag().contains("effect_" + i)) {
-//                    MobEffect effect = MobEffect.byId(i);
-//                    if (effect != null) {
-//                        MobEffectInstance instance = new MobEffectInstance(effect, (int) NBTUtils.getFloatOrAddKey(stack, "effect_" + i) * 20, 1, false, false);
-//                        builder.effect(() -> instance, 1);
-//                    }
-//                }
-//            }
-//        }
+        List<EffectData> effects = stack.getOrDefault(DataComponents.EFFECTS, Collections.emptyList());
+
+        for (EffectData effect : effects) {
+            Optional<Holder.Reference<MobEffect>> mobEffect = BuiltInRegistries.MOB_EFFECT.getHolder(ResourceLocation.withDefaultNamespace(effect.resourceLocation()));
+            if (mobEffect.isPresent()) {
+                MobEffectInstance instance = new MobEffectInstance(mobEffect.get(), effect.duration() * 20, 1, false, false);
+                builder.effect(() -> instance, 1);
+            }
+        }
 
         return builder.build();
     }
-
 
     @Override
     public InteractionResultHolder<ItemStack> use(Level level, Player player, InteractionHand usedHand) {
@@ -123,5 +128,10 @@ public class CrystalApple extends LevelableTool {
         }
 
         return true;
+    }
+
+    @Override
+    public UseAnim getUseAnimation(ItemStack pStack) {
+        return UseAnim.EAT;
     }
 }